+++ src/package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: 'xolvio:md-blog',
   summary: 'A markdown powered blog',
-  version: '0.3.9',
+  version: '0.4.0',
   git: 'https://github.com/xolvio/md-blog'
 });